chore: 解决启动问题

+++ spring-loopnurture-users-service/src/main/java/org/springframework/samples/loopnurture/users/infra/mapper/JpaMarketingUserMapper.java
@@ -16,5 +16,5 @@ public interface JpaMarketingUserMapper extends JpaRepository<MarketingUserPO, S
     boolean existsByUserUniq(String userUniq);
     boolean existsByPrimaryEmail(String primaryEmail);
     boolean existsByPhone(String phone);
-    Optional<MarketingUserPO> findByOauthUserIdAndAuthType(String oauthUserId, Integer authType);
+    Optional<MarketingUserPO> findByOauthUserIdAndAuthType(String oauthUserId, Short authType);
 }
